requirements.txt
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
dirtyjson==1.0.8
diskcache==5.6.3
distro==1.9.0
environs==9.5.0
exceptiongroup==1.2.1
filelock==3.14.0
fsspec==2024.5.0
grpcio==1.63.0
h11==0.14.0
httpcore==1.0.5
httpx==0.25.2
huggingface-hub==0.23.0
idna==3.7
iniconfig==2.0.0
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
llama_cpp_python==0.2.11
MarkupSafe==2.1.5
marshmallow==3.21.2
milvus-lite==2.4.4
mistralai==0.1.8
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
openai==1.30.1
orjson==3.10.3
packaging==24.0
pandas==2.2.2
pillow==10.3.0
pluggy==1.5.0
protobuf==5.26.1
py-llm-core==2.8.4
pydantic==2.7.1
pydantic_core==2.18.2
pymilvus==2.4.3
pytest==8.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
regex==2024.5.15
requests==2.31.0
rpds-py==0.18.1
safetensors==0.4.3
scikit-learn==1.4.2
scipy==1.13.0
sentence-transformers==2.7.0
six==1.16.0
sniffio==1.3.1
sympy==1.12
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.19.1
tomli==2.0.1
torch==2.3.0
tqdm==4.66.4
transformers==4.41.0
typing_extensions==4.11.0
tzdata==2024.1
ujson==5.10.0
urllib3==2.2.1
nltk==3.9.1
